#9 Controllers Changed and New Script Written.
