+++ rover_21_robotic_arm/arm_21_gazebo/urdf/arm_21_gazebo.urdf.xacro
@@ -387,7 +387,7 @@
   </link>
   <joint
     name="joint_6"
-    type="revolute">
+    type="continuous">
     <origin
       xyz="0.00949310528755105 0.17173782621193 -0.0277500000000037"
       rpy="-1.5707963267949 0.209223763052682 -0.0552205323847502" />
