+++ rover_21_robotic_arm/arm_control/config/arm21_controller.yaml
@@ -1,65 +1,45 @@
+# Simulation settings for using moveit_sim_controllers
+moveit_sim_hw_interface:
+  joint_model_group: manipulator
+  joint_model_group_pose: home
+# Settings for ros_control_boilerplate control loop
+generic_hw_control_loop:
+  loop_hz: 300
+  cycle_time_error_threshold: 0.01
+# Settings for ros_control hardware interface
+hardware_interface:
+  joints:
+    - joint_1
+    - joint_2
+    - joint_3
+    - joint_4
+    - joint_5
+    - joint_6
+    - joint_left
+    - joint_right
+    - left_front_joint
+    - left_rear_joint
+    - right_front_joint
+    - right_rear_joint
+  sim_control_mode: 1  # 0: position, 1: velocity
+controller_list:
+  []
 rover_arm_controller:
   type: position_controllers/JointTrajectoryController
   joints:
-     - joint_1
-     - joint_2
-     - joint_3
-     - joint_4
-     - joint_5
-     - joint_6
-  constraints:
-      goal_time: 0.6
-      stopped_velocity_tolerance: 0.05
-      joint_1: {trajectory: 0.1, goal: 0.1}
-      joint_2: {trajectory: 0.1, goal: 0.1}
-      joint_3: {trajectory: 0.1, goal: 0.1}
-      joint_4: {trajectory: 0.1, goal: 0.1}
-      joint_5: {trajectory: 0.1, goal: 0.1}
-      joint_6: {trajectory: 0.1, goal: 0.1}
-  stop_trajectory_duration: 0.5
-  state_publish_rate:  25
-  action_monitor_rate: 10
+    - joint_1
+    - joint_2
+    - joint_3
+    - joint_4
+    - joint_5
+    - joint_6
 
-
-
-#buranin asagisindan emin degilim, deneme asamasinda
 
 rover_arm_right_finger:
     type: position_controllers/JointPositionController
     joint: joint_right
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
+
 
 rover_arm_left_finger:
     type: position_controllers/JointPositionController
     joint: joint_left
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
-
-rover_arm_joint_1:
-    type: position_controllers/JointPositionController
-    joint: joint_1
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
-    
-rover_arm_joint_2:
-    type: position_controllers/JointPositionController
-    joint: joint_2
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
-
-rover_arm_joint_3:
-    type: position_controllers/JointPositionController
-    joint: joint_3
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
-
-rover_arm_joint_4:
-    type: position_controllers/JointPositionController
-    joint: joint_4
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
-
-rover_arm_joint_5:
-    type: position_controllers/JointPositionController
-    joint: joint_5
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
-
-rover_arm_joint_6:
-    type: position_controllers/JointPositionController
-    joint: joint_6
-    pid: {p: 1000,  d: 10, i: 0.1, i_clamp: 1}
