+++ rover_21_robotic_arm/robot_drive/launch/kinematic.launch
@@ -11,5 +11,5 @@
     <param name="kill" value="4" type="int" />
     <param name="turbo" value="5" type="int" />
     <param name="turbo_scale" value="2" type="double" />
-    <node pkg="robot_drive" name="forward_kinematic" type="forward_kinematic.py" output="screen" /> <!-- -->
+    <node pkg="robot_drive" name="kinematic" type="kinematic.py" output="screen" /> <!-- -->
 </launch>
