--- rover_21_robotic_arm/robot_drive/src/Fabrik/coordinate.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-import math
-
-
-
-
-
-
